+++ components/disclosure/circle-detail-tabs.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import type { AlertStatus, FC } from "@yamada-ui/react"
+import type { FC } from "@yamada-ui/react"
 import {
   Indicator,
   Avatar,
@@ -12,7 +12,6 @@ import {
   MultiSelect,
   Option,
   SimpleGrid,
-  Snacks,
   Tab,
   TabList,
   TabPanel,
@@ -21,19 +20,13 @@ import {
   Tag,
   Text,
   VStack,
-  useSnacks,
 } from "@yamada-ui/react"
 import Link from "next/link"
 import { useState } from "react"
 import { CircleActivitydays } from "../data-display/circle-activitydays"
-import { MemberCard } from "../data-display/member-card"
-import { MemberRequestCard } from "../data-display/member-request-card"
+import { CircleMembers } from "../data-display/circle-members"
 import type { getCircleById } from "@/actions/circle/fetch-circle"
-import {
-  removeMember,
-  type getMembershipRequests,
-} from "@/actions/circle/membership-request"
-import { changeMemberRole } from "@/actions/circle/update-role"
+import { type getMembershipRequests } from "@/actions/circle/membership-request"
 import type { getActivityById } from "@/data/activity"
 
 interface CircleDetailTabsProps {
@@ -116,56 +109,6 @@ export const CircleDetailTabs: FC<CircleDetailTabsProps> = ({
   const userRole = circle?.members?.find((member) => member.id === userId)?.role
   const tabIndex = handlingTab(tabKey || "")
   const { data } = membershipRequests
-  const { snack, snacks } = useSnacks()
-  const handleSnack = (title: string, status: AlertStatus) => {
-    snack.closeAll()
-    snack({ title, status })
-  }
-  const handleRoleChange = async (
-    targetMemberId: string,
-    newRoleId: number,
-  ) => {
-    try {
-      const { message, success } = await changeMemberRole({
-        userId, // 現在のユーザーID
-        circleId: circle?.id || "", // サークルID
-        targetMemberId, // 変更対象のメンバーID
-        newRoleId, // 新しい役職ID
-      })
-      if (success) {
-        handleSnack(message, "success")
-        await fetchData() // データを再フェッチ
-      } else {
-        handleSnack(message, "error")
-      }
-    } catch (error) {
-      handleSnack(
-        error instanceof Error ? error.message : "エラーが発生しました。",
-        "error",
-      )
-    }
-  }
-
-  const handleRemoveMember = async (targetMemberId: string) => {
-    try {
-      const { message, success } = await removeMember({
-        circleId: circle?.id || "",
-        targetMemberId,
-        userId,
-      })
-      if (success) {
-        handleSnack(message, "success")
-        await fetchData() // データを再フェッチ
-      } else {
-        handleSnack(message, "error")
-      }
-    } catch (error) {
-      handleSnack(
-        error instanceof Error ? error.message : "エラーが発生しました。",
-        "error",
-      )
-    }
-  }
 
   return (
     <Tabs index={tabIndex}>
@@ -267,9 +210,7 @@ export const CircleDetailTabs: FC<CircleDetailTabsProps> = ({
                         )}
                         <Text fontWeight="bold">{item.title}</Text>
                       </HStack>
-                      <VStack
-                        align="end"
-                      >
+                      <VStack align="end">
                         <Text fontSize="sm" color="gray.500">
                           {item.date} {item.time}
                         </Text>
@@ -282,30 +223,14 @@ export const CircleDetailTabs: FC<CircleDetailTabsProps> = ({
         </TabPanel>
 
         <TabPanel>
-          <Snacks snacks={snacks} />
-          <SimpleGrid w="full" columns={{ base: 2, md: 1 }} gap="md">
-            {data?.map((member) => (
-              <MemberRequestCard
-                key={member.id}
-                member={member}
-                userId={userId}
-                circleId={circle?.id || ""}
-                fetchData={fetchData}
-                handleSnack={handleSnack}
-              />
-            ))}
-            {circle?.members?.map((member) => (
-              <MemberCard
-                key={member.id}
-                member={member}
-                isAdmin={isAdmin}
-                userId={userId}
-                userRole={userRole}
-                handleRoleChange={handleRoleChange}
-                handleRemoveMember={handleRemoveMember}
-              />
-            ))}
-          </SimpleGrid>
+          <CircleMembers
+            userId={userId}
+            userRole={userRole}
+            isAdmin={!!isAdmin}
+            circle={circle}
+            membershipRequests={membershipRequests}
+            fetchData={fetchData}
+          />
         </TabPanel>
       </TabPanels>
     </Tabs>
